backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo
pymongo[srv]
python-dotenv
httpx
cryptography
supabase
